pyproject.toml
[tool.poetry]
name = "pythonproject unit-5"
version = "0.1.0"
description = ""
authors = ["Alexandra Goldobina <goldobina.s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
response = "^0.5.0"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
